requirements.txt
Django==3.2.5
djangorestframework==3.13.1
djangorestframework_simplejwt==5.1.0
